# Faster Q4_K_R4 and Q5_K_R4 on AVX2/Zen4